@@ -16,6 +16,7 @@ package org.eclipse.jetty.server.handler;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
+import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.handler.ContextHandler.AliasCheck;
 import org.eclipse.jetty.util.resource.PathResource;
 import org.eclipse.jetty.util.resource.Resource;
@@ -24,9 +25,10 @@ import org.slf4j.LoggerFactory;
 
 /**
  * Symbolic Link AliasChecker.
- * <p>An instance of this class can be registered with {@link ContextHandler#addAliasCheck(AliasCheck)}
- * to check resources that are aliased to other locations.   The checker uses the
- * Java {@link Files#readSymbolicLink(Path)} and {@link Path#toRealPath(java.nio.file.LinkOption...)}
+ * <p>An instance of this class can be registered with {@link Handler} subclasses
+ * to check resources that are aliased to other locations.</p>
+ * <p>The checker uses the Java {@link Files#readSymbolicLink(Path)} and
+ * {@link Path#toRealPath(java.nio.file.LinkOption...)}
  * APIs to check if a file is aliased with symbolic links.</p>
  * @deprecated use {@link org.eclipse.jetty.server.SymlinkAllowedResourceAliasChecker} instead.
  */
