@@ -13,23 +13,8 @@
 
 package org.eclipse.jetty.ee10.demos;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.Map;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.eclipse.jetty.server.CustomRequestLog;
 import org.eclipse.jetty.server.Handler;
-import org.eclipse.jetty.server.Request;
-import org.eclipse.jetty.server.Server;
-import org.eclipse.jetty.server.handler.ContextHandler;
-import org.eclipse.jetty.server.handler.ContextHandlerCollection;
 import org.eclipse.jetty.server.handler.DefaultHandler;
-import org.eclipse.jetty.server.handler.gzip.GzipHandler;
-import org.eclipse.jetty.util.ajax.JSON;
 
 /**
  * Frequently many handlers are combined together to handle different aspects of
@@ -43,16 +28,16 @@ import org.eclipse.jetty.util.ajax.JSON;
  * </ul>
  * Multiple handlers may be combined with:
  * <ul>
- * <li>{@link HandlerWrapper} which will nest one handler inside another. In
+ * <li>{@link Handler.Wrapper} which will nest one handler inside another. In
  * this example, the HelloHandler is nested inside a HandlerWrapper that sets
  * the greeting as a request attribute.
- * <li>{@link HandlerList} which will call a collection of handlers until the
+ * <li>{@link Handler.Collection} which will call a collection of handlers until the
  * request is marked as handled. In this example, a list is used to combine the
  * param handler (which only handles the request if there are parameters) and
  * the wrapper handler. Frequently handler lists are terminated with the
  * {@link DefaultHandler}, which will generate a suitable 404 response if the
  * request has not been handled.
- * <li>{@link HandlerCollection} which will call each handler regardless if the
+ * <li>{@link Handler.Collection} which will call each handler regardless if the
  * request has been handled or not. Typically this is used to always pass a
  * request to the logging handler.
  * </ul>
